moved addSubmission into /db/helper.js
